+++ main/idf_component.yml
@@ -1,4 +1,5 @@
 ## IDF Component Manager Manifest File
 dependencies:
-  espressif/esp_tinyusb: "^2.0.0"
-  idf: ">=5.1"
+  espressif/esp_tinyusb:
+    version: "^2.0.0"
+  idf: "^5.0"
